CLIPModel: fix typing

+++ src/lib/clip/CLIPModel.py
@@ -29,5 +29,5 @@ class CLIPModel(torch.nn.Module):
 		
 		# result = torch.adaptive_avg_pool1d(result, 1).squeeze(-1)
 		result = result.flatten(start_dim=1) # [ batch_size, 1024 ]
-		result = self.classifier(result) # [ batch_size, 2 ]
+		result = self.classifier(result.type(torch.FloatTensor)) # [ batch_size, 2 ]
 		return result
